== Attendance System == =General: *Removed some duplicated clif_functions. -Fixed some other functions and did some clean-up.
*Changed a huge number of functions to support int64 values.
-Needed to probably handle timestamps and more...

*Implementation of Attendance system.
-added battle/attendance.conf configuration file.
-Make sure to also edit your system/CheckAttendance.lub file with the same values as you added server-sided for this to be displayed probably ingame.
-A 2018-06-21aRagexeRE will be required for this feature to work.

+++ src/map/mapreg.h
@@ -9,9 +9,9 @@ void mapreg_final(void);
 void mapreg_init(void);
 bool mapreg_config_read(const char* w1, const char* w2);
 
-int mapreg_readreg(int uid);
-char* mapreg_readregstr(int uid);
-bool mapreg_setreg(int uid, int val);
-bool mapreg_setregstr(int uid, const char* str);
+int64 mapreg_readreg(int64 uid);
+char* mapreg_readregstr(int64 uid);
+bool mapreg_setreg(int64 uid, int64 val);
+bool mapreg_setregstr(int64 uid, const char* str);
 
 #endif /* _MAPREG_H_ */
